@@ -36,11 +36,9 @@
   <!-- Shared properties -->
   <properties>
     <!-- Frameworks -->
-    <java.version>11</java.version>
-    <java.version_max>16</java.version_max>
-    <java.version_end>17</java.version_end>
-    <maven.version>3.6.0</maven.version>
-    <spring.version>5.3.20</spring.version>
+    <java.version>21</java.version>
+    <maven.version>3.8.8</maven.version>
+    <spring.version>6.2.0</spring.version>
 
     <!-- Build layout -->
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
